refactor(routing): extract hint matching helpers
